PowerStatsService: Prevent NPE in BatteryTrigger.
The ACTION_BATTERY_CHANGED Broadcast is a sticky broadcast, and
the registerReceiver method will return null when the first sticky
intent found that not matches filter.
However, the constructor of BatteryTrigger does not check whether
batteryStatus is null and uses it directly. In SystemServer,
PowerStatsService is started earlier than BatteryService.
If SystemServer is blocked for a few seconds for some reason,
then the NullPointerException will be thrown by BatteryTrigger.

Bug: N/A
Test: atest FrameworksServicesTests:PowerStatsServiceTest
Change-Id: I0e29e4a2c7e1298065d8b43c3b4b2bf2afb975e1